// src/services/models/ServiceViewPagedCollection.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Link } from './Link';
import type { ServiceView } from './ServiceView';
export type ServiceViewPagedCollection = {
  offset?: number | null;
  limit?: number | null;
  size?: number;
  first?: Link;
  previous?: Link;
  next?: Link;
  last?: Link;
  self?: Link;
  value?: Array<ServiceView> | null;
  nextOffset?: number | null;
  previousOffset?: number | null;
};